separate base schema from MoH-specific schema
